# docs/api_spec_full.yaml
openapi: 3.0.3
info:
  title: SoulRep Gym Management API
  description: |
    Complete API suite for SoulRep Gym.
    Covers Auth, Member Dashboard, Owner Analytics, and Trainer Operations.
  version: 2.1.0
servers:
  - url: http://localhost:3000/api
paths:
...
  /signup:
    post:
      summary: Register new member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string }
                email: { type: string }
                phone: { type: string }
                password: { type: string }
      responses:
        '201': { description: Registered }
...
  /trainer/members/{memberId}/workout-plans:
    get:
      summary: Get workout plans for specific member
      parameters:
        - name: memberId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Plans }
    post:
      summary: Create workout plan for member
      parameters:
        - name: memberId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [day, exercises]
              properties:
                day: { type: string }
                notes: { type: string }
                exercises:
                  type: array
                  items:
                    type: object
                    required: [name, sets, reps, rest]
                    properties:
                      name: { type: string }
                      sets: { type: integer }
                      reps: { type: string }
                      rest: { type: string }
                      notes: { type: string }
      responses:
        '201': { description: Created }
...